backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["seiwa-ic-system <s-system@seiwa-ic.co.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.8.2"}
python-multipart = "^0.0.5"
alembic = "^1.7.5"
SQLAlchemy = "^1.4.27"
psycopg2-binary = "^2.9.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.910"
black = "^21.11b1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
